Fix CPU fan delay on temperature increase (#10)
My Dell XPS 9360 has been overheating for awhile now. Since it is a 5
year old laptop, I've _assumed_ (uh uh...) that the ~40 degrees
temperatures of where I'm living, plus the dust likely collected by the
fan were the issue.

For that reason, I've started by cleaning up the fan, replacing the
thermal past on the CPU, which unfortunately did not made that much of
an improvement.

I've then took the laptop to a local computer shop, where the person
there told me that, by the sound of it, the fan should be on the last of
its days. This is when I found out that Dell does not resell these
components (at least not for 5 year old laptops...), so I had to dig a
bit on the internet to find and purchase replacement. After installing
that new fan, while it was a bit better (i.e., it didn't overheat so
easily), it was still a problem...

Then I finally did what I should have done from the start: pay attention
to what is happening to the CPU temperatures and to the fan when I
increase utilisation on the laptop. It seemed to me then that the issue is
the fan taking several seconds to spin up properly after the CPU
gets to 80+ degrees, which then leads to the fan not being able to
cooldown the CPU fast enough.

I then read about Dell's System Management Mode (SSM) BIOS, that
basically overtakes fan control from the operating system. [1]
Seems like this is the sole reason for [`dell-smm-hwmon` kernel module](https://docs.kernel.org/hwmon/dell-smm-hwmon.html#description)'s
existence, but unfortunately,
[it doesn't know a thing about this particular
laptop](https://github.com/torvalds/linux/blob/2dde18cd1d8fac735875f2e4987f11817cc0bc2c/drivers/hwmon/dell-smm-hwmon.c#L1292C20-L1292C20)
[2], meaning that I had find another way
retake control of the fan to the OS.
Finally, [the Archlinux wiki article on
this](https://wiki.archlinux.org/title/Fan_speed_control#Dell_laptops)
mentions a fork of a `dell-ssm-hwmon`-related tool that does handle this
laptop's Dell BIOS properly, so that it can disable BIOS fan control.

So with that all out of the way, we can setup `fancontrol` to properly
keep the temperature low on the CPU, by running `pwmconfig` to generate
its configuration, enabling the `fancontrol` service, and making sure it
is running on sleep resume. [3]

1: Coincidentely, I did update the BIOS firmware around the time that
   this whole ordeal started, but I cannot for the life of me remember
   if it was before or after the first overheating incidents.
2: I'm sure not to anyone's fault but Dell, for not providing these codes to the kernel developers.
3: It seems to be a known issue that the daemon will stop working on
   resume: https://github.com/lm-sensors/lm-sensors/issues/172